сделала правки
